requirements-build.txt
PyYAML ~= 5.3.1
